requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
torch
pdfplumber
camelot-py
opencv-python
ghostscript
opencv-python
ghostscript
PyPDF2
pytesseract
pdf2image
Pillow
beautifulsoup4
requests
lxml
mysql-connector-python
jsonschema
numpy
pandas
pytest
httpx
playwright
tiktoken 
asyncio